DF-7876 cast to int
Former-commit-id: aa5112dedd6f6f26ca0394ba559ca2462a7d818a

+++ Sources/Tools/TexturePacker/ResourcePacker2D.cpp
@@ -72,7 +72,7 @@ const String& GetCodeAsString(AssetClientCode code)
                                             "CANNOT_READ_FILES"}};
     static String codeUnknown("CODE_UNKNOWN");
 
-    if (code >= 0 && code < codeStrings.size())
+    if (code >= 0 && code < static_cast<int>(codeStrings.size()))
     {
         return codeStrings[code];
     }
